+++ src/app/page.tsx
@@ -1,73 +1,32 @@
 "use client";
 import { AttestationItem } from "@/components/AttestationItem";
-import { CREATE_EVENT_SCHEMA_UID } from "@/hooks/useCreateEventAttestation";
-import { useSigner } from "@/hooks/useSigner";
-import { API_URL_MAPPING } from "@/lib/gqlEasAttestation";
-import { EVENTS_ATTESTATIONS_QUERY } from "@/lib/gqlEasAttestation/query";
 import { useMemo } from "react";
-import { useQuery } from "urql";
-import { useChainId } from "wagmi";
-
-interface DataEntry {
-  attester: string;
-  data: string;
-  decodedDataJson: string;
-  expirationTime: number;
-  id: string;
-  recipient: string;
-  refUID: string;
-  revocable: boolean;
-  revocationTime: number;
-}
-
-function sortByStartsAt(data: DataEntry[]): DataEntry[] {
-  return data.sort((a, b) => {
-    const aDecoded = JSON.parse(a.decodedDataJson);
-    const bDecoded = JSON.parse(b.decodedDataJson);
-
-    const aStartsAt = aDecoded.find((item: any) => item.name === "startsAt")
-      ?.value.value.hex;
-    const bStartsAt = bDecoded.find((item: any) => item.name === "startsAt")
-      ?.value.value.hex;
-
-    if (!aStartsAt || !bStartsAt) return 0;
-
-    return Number.parseInt(aStartsAt, 16) - Number.parseInt(bStartsAt, 16);
-  });
-}
+import {
+  type DataEntry,
+  useAllEventsData,
+} from "./event/hooks/useAllEventsData";
 
 export default function Events() {
-  const signer = useSigner();
-
-  const chainId = useChainId();
-
-  const [result] = useQuery({
-    query: EVENTS_ATTESTATIONS_QUERY,
-    variables: { schemaId: CREATE_EVENT_SCHEMA_UID },
-    context: useMemo(() => ({ url: API_URL_MAPPING[chainId] }), [chainId]),
-    pause: !signer,
-  });
-
-  const { data } = result;
+  const { allEventsData, signer } = useAllEventsData();
 
   const attestationList = useMemo(
     () =>
-      data?.attestations &&
-      sortByStartsAt(data.attestations).map((attestation) => (
+      allEventsData &&
+      allEventsData.map((attestation: DataEntry) => (
         <AttestationItem
           id={attestation.id}
           key={attestation.id}
           data={attestation.decodedDataJson}
         />
       )),
-    [data?.attestations],
+    [allEventsData],
   );
 
   if (!signer) {
     return <div>Connect a wallet to view your attestations.</div>;
   }
 
-  if (!data) {
+  if (!allEventsData) {
     return <div>Loading attestation data...</div>;
   }
